feat(zsh): read in .zshrc.d files
This adds support for `zsh` to read in all the files from `~/.zshrc.d`
and load the at the start of each shell. This makes it a lot easier for
individual applications to add configuration settings without having to
add them to the master `.zshrc` file.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>